+++ app/Helpers/AuthenticateByTokenService.ts
